Optimization: Added option to give in the "parentEntityType" for saving and creating Items. When given, the GCL/ICL checks for the linktype. If the linktype is set to UserParentId then it saves addding a link and saves the parent imediately, this saves a few query's.
Also: changed the order of getting if a link is ordened by 'order' or by 'alphabet', in case a link between items is ordered by alphabet, we don't need to check for the new 'ordernumber'. This also saves a query per creation of an item. For example basketlines in a basket. We also save this information on creation instead of doing this afterwards in an update (this was not very efficient). Because of this now we always load the linktype, but this query is cached and therefore this way is more efficient alltogether.

Also: on saving an item, if the item is newly created, there is no need to check if the item exists, as it has just been created so it always exists. This saves an unnecessary query at creation of an item.

Feature: at creation of an item save the addedOn date if given in the object.

Feature: at creation of an item, if the AddedBy property is empty then use the username, otherwise use the AddedBy property

Set the Changed property to false for the item at creation, because the item is then always new and so not changed.